+++ plugins/Kaleidoscope-LED-ActiveModColor/src/kaleidoscope/plugin/LED-ActiveModColor.cpp
@@ -17,6 +17,7 @@
 
 #include <Kaleidoscope-LED-ActiveModColor.h>
 #include <Kaleidoscope-OneShot.h>
+#include <Kaleidoscope-OneShotMetaKeys.h>
 #include "kaleidoscope/LiveKeys.h"
 #include "kaleidoscope/layers.h"
 #include "kaleidoscope/keyswitch_state.h"
@@ -54,7 +55,7 @@ EventHandlerResult ActiveModColorEffect::onKeyEvent(KeyEvent &event) {
         // Get the entry from the live keys array
         Key entry_key = live_keys[entry_addr];
         // Skip empty entries
-        if (entry_key == Key_Transparent || entry_key == Key_NoKey) {
+        if (entry_key == Key_Inactive || entry_key == Key_Masked) {
           continue;
         }
         // Highlight everything else