+++ plugins/Kaleidoscope-OneShot/src/kaleidoscope/plugin/OneShot.cpp
@@ -52,10 +52,6 @@ KeyAddrBitfield OneShot::glue_addrs_;
 uint16_t OneShot::start_time_ = 0;
 KeyAddr OneShot::prev_key_addr_ = OneShot::invalid_key_addr;
 
-#ifndef ONESHOT_WITHOUT_METASTICKY
-KeyAddr OneShot::meta_sticky_key_addr_ {KeyAddr::invalid_state};
-#endif
-
 
 // ============================================================================
 // Public interface
@@ -118,26 +114,38 @@ bool OneShot::isSticky() {
 // could potentially use three different color values for the three
 // states (sticky | active && !sticky | pressed && !active).
 
+__attribute__((weak))
 bool OneShot::isStickable(Key key) {
+  return isStickableDefault(key);
+}
+
+bool OneShot::isStickableDefault(Key key) {
   int8_t n;
+  // If the key is either a keyboard modifier or a layer shift, we check to see
+  // if it has been set to be non-stickable.
   if (key.isKeyboardModifier()) {
     n = key.getKeyCode() - Key_LeftControl.getKeyCode();
     return bitRead(stickable_keys_, n);
   } else if (key.isLayerShift()) {
     n = oneshot_mod_count + key.getKeyCode() - LAYER_SHIFT_OFFSET;
+    // We only keep track of the stickability of the first 8 layers.
     if (n < oneshot_key_count) {
       return bitRead(stickable_keys_, n);
     }
-#ifndef ONESHOT_WITHOUT_METASTICKY
-  } else if (key == OneShot_MetaStickyKey) {
-    return true;
-#endif
   }
-  return false;
+  // The default is for all keys to be "stickable"; if the default was false,
+  // any user code or other plugin that uses `setPending()` to turn a key into a
+  // OneShot would need to override `isStickable()` in order to make that key
+  // stickable (the default `OSM()` behaviour).
+  return true;
 }
 
 bool OneShot::isTemporary(KeyAddr key_addr) {
   return temp_addrs_.read(key_addr);
+}
+
+bool OneShot::isPending(KeyAddr key_addr) {
+  return (glue_addrs_.read(key_addr) && temp_addrs_.read(key_addr));
 }
 
 bool OneShot::isSticky(KeyAddr key_addr) {
@@ -146,6 +154,31 @@ bool OneShot::isSticky(KeyAddr key_addr) {
 
 bool OneShot::isActive(KeyAddr key_addr) {
   return (isTemporary(key_addr) || glue_addrs_.read(key_addr));
+}
+
+// ----------------------------------------------------------------------------
+// Public state-setting functions
+
+void OneShot::setPending(KeyAddr key_addr) {
+  temp_addrs_.set(key_addr);
+  glue_addrs_.clear(key_addr);
+  start_time_ = Runtime.millisAtCycleStart();
+}
+
+void OneShot::setOneShot(KeyAddr key_addr) {
+  temp_addrs_.set(key_addr);
+  glue_addrs_.set(key_addr);
+  start_time_ = Runtime.millisAtCycleStart();
+}
+
+void OneShot::setSticky(KeyAddr key_addr) {
+  temp_addrs_.clear(key_addr);
+  glue_addrs_.set(key_addr);
+}
+
+void OneShot::clear(KeyAddr key_addr) {
+  temp_addrs_.clear(key_addr);
+  glue_addrs_.clear(key_addr);
 }
 
 // ----------------------------------------------------------------------------
@@ -190,27 +223,6 @@ EventHandlerResult OneShot::onKeyEvent(KeyEvent& event) {
 
   if (keyToggledOn(event.state)) {
 
-    // Make all held keys sticky if `OneShot_ActiveStickyKey` toggles on.
-    if (event.key == OneShot_ActiveStickyKey) {
-      // Skip the stickify key itself
-      for (KeyAddr entry_addr : KeyAddr::all()) {
-        if (entry_addr == event.addr) {
-          continue;
-        }
-        // Get the entry from the keyboard state array
-        Key entry_key = live_keys[entry_addr];
-        // Skip empty entries
-        if (entry_key == Key_Transparent || entry_key == Key_NoKey) {
-          continue;
-        }
-        // Make everything else sticky
-        temp_addrs_.clear(entry_addr);
-        glue_addrs_.set(entry_addr);
-      }
-      prev_key_addr_ = event.addr;
-      return EventHandlerResult::OK;
-    }
-
     if (!temp && !glue) {
       // The key is in the "normal" state. The first thing we need to do is
       // convert OneShot keys to their equivalent values, and record the fact
@@ -221,34 +233,6 @@ EventHandlerResult OneShot::onKeyEvent(KeyEvent& event) {
         event.key = decodeOneShotKey(event.key);
         is_oneshot = true;
       }
-
-#ifndef ONESHOT_WITHOUT_METASTICKY
-      bool is_meta_sticky_key_active = meta_sticky_key_addr_.isValid();
-      if (is_meta_sticky_key_active) {
-        // If the meta key isn't sticky, release it
-        bool ms_temp = temp_addrs_.read(meta_sticky_key_addr_);
-        bool ms_glue = glue_addrs_.read(meta_sticky_key_addr_);
-        if (ms_temp) {
-          if (ms_glue) {
-            // The meta key is in the "one-shot" state; release it immediately.
-            releaseKey(meta_sticky_key_addr_);
-          } else {
-            // The meta key is in the "pending" state; cancel that, and let it
-            // deactivate on release.
-            temp_addrs_.clear(meta_sticky_key_addr_);
-          }
-        }
-        glue_addrs_.set(event.addr);
-      } else if (event.key == OneShot_MetaStickyKey) {
-        meta_sticky_key_addr_ = event.addr;
-        temp_addrs_.set(event.addr);
-      }
-      if (is_meta_sticky_key_active || (event.key == OneShot_MetaStickyKey)) {
-        prev_key_addr_ = event.addr;
-        start_time_ = Runtime.millisAtCycleStart();
-        return EventHandlerResult::OK;
-      }
-#endif
 
       if (is_oneshot ||
           (auto_modifiers_ && event.key.isKeyboardModifier()) ||
@@ -309,11 +293,6 @@ EventHandlerResult OneShot::onKeyEvent(KeyEvent& event) {
       // stop that event from sending a report, and instead send a "hold"
       // event. This is handled in the `beforeReportingState()` hook below.
       return EventHandlerResult::ABORT;
-#ifndef ONESHOT_WITHOUT_METASTICKY
-    } else if (event.key == OneShot_MetaStickyKey) {
-      // Turn off the meta key if it's released in its "normal" state.
-      meta_sticky_key_addr_ = KeyAddr::none();
-#endif
     }
 
   }
@@ -433,11 +412,6 @@ void OneShot::holdKey(KeyAddr key_addr) {
 void OneShot::releaseKey(KeyAddr key_addr) {
   glue_addrs_.clear(key_addr);
   temp_addrs_.clear(key_addr);
-
-#ifndef ONESHOT_WITHOUT_METASTICKY
-  if (live_keys[key_addr] == OneShot_MetaStickyKey)
-    meta_sticky_key_addr_ = KeyAddr::none();
-#endif
 
   KeyEvent event{key_addr, WAS_PRESSED | INJECTED};
   Runtime.handleKeyEvent(event);